requirements.txt
flask>=2.0.0
requests>=2.28.0
auggie-sdk>=0.1.10
pymongo>=4.0.0

# Slack integration (optional)
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# AI Middleware dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-multipart>=0.0.6
aiofiles>=23.2.1
tenacity>=8.2.3
structlog>=24.1.0
python-dotenv>=1.0.0

